Only parse code for assert lib on demand
