+++ Pluto.vcxproj
@@ -418,6 +418,7 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="src\lapi.cpp" />
+    <ClCompile Include="src\lassertlib.cpp" />
     <ClCompile Include="src\lauxlib.cpp" />
     <ClCompile Include="src\lbase32.cpp" />
     <ClCompile Include="src\lbase64.cpp" />
